@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,6 +27,8 @@ dependencies:
   youtube_player_flutter: ^8.0.0
   expandable: ^5.0.1
   get: ^4.1.4
+  hexcolor: ^2.0.4
+  
 
 dev_dependencies:
   flutter_test:
